src/tf-v1.15/requirements.txt
matplotlib
pandas
numpy
sklearn
tensorflow-gpu==1.15

